Refactor roles utilities
This brings all basic roles utilities into one file, separating them
from other utilities used to display checkboxes or build lists of select
options.
